+++ x/gov/types/v1beta1/codec.go
@@ -4,7 +4,6 @@ import (
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/msgservice"
-	distrtypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
 )
 
 // RegisterInterfaces registers the interfaces types with the Interface Registry.
@@ -19,10 +18,6 @@ func RegisterInterfaces(registry codectypes.InterfaceRegistry) {
 		"cosmos.gov.v1beta1.Content",
 		(*Content)(nil),
 		&TextProposal{},
-	)
-	registry.RegisterImplementations(
-		(*Content)(nil),
-		&distrtypes.CommunityPoolSpendProposal{}, //nolint: staticcheck // avoid using `CommunityPoolSpendProposal`, might be reomved in future.
 	)
 
 	msgservice.RegisterMsgServiceDesc(registry, &_Msg_serviceDesc)