+++ x/slashing/abci_test.go
@@ -4,10 +4,9 @@ import (
 	"testing"
 	"time"
 
-	abci "github.com/cometbft/cometbft/abci/types"
-	cmtproto "github.com/cometbft/cometbft/proto/tendermint/types"
 	"github.com/stretchr/testify/require"
 
+	"cosmossdk.io/core/comet"
 	"cosmossdk.io/depinject"
 	"cosmossdk.io/log"
 
@@ -65,15 +64,17 @@ func TestBeginBlocker(t *testing.T) {
 	require.NoError(t, err)
 	require.Equal(t, amt, val.GetBondedTokens())
 
-	abciVal := abci.Validator{
+	abciVal := comet.Validator{
 		Address: pk.Address(),
 		Power:   power,
 	}
 
-	ctx = ctx.WithVoteInfos([]abci.VoteInfo{{
-		Validator:   abciVal,
-		BlockIdFlag: cmtproto.BlockIDFlagCommit,
-	}})
+	ctx = ctx.WithCometInfo(comet.Info{
+		LastCommit: comet.CommitInfo{Votes: []comet.VoteInfo{{
+			Validator:   abciVal,
+			BlockIDFlag: comet.BlockIDFlagCommit,
+		}}},
+	})
 
 	err = slashing.BeginBlocker(ctx, slashingKeeper)
 	require.NoError(t, err)
@@ -91,11 +92,7 @@ func TestBeginBlocker(t *testing.T) {
 	require.NoError(t, err)
 	// for 100 blocks, mark the validator as having signed
 	for ; height < signedBlocksWindow; height++ {
-		ctx = ctx.WithBlockHeight(height).
-			WithVoteInfos([]abci.VoteInfo{{
-				Validator:   abciVal,
-				BlockIdFlag: cmtproto.BlockIDFlagCommit,
-			}})
+		ctx = ctx.WithBlockHeight(height)
 
 		err = slashing.BeginBlocker(ctx, slashingKeeper)
 		require.NoError(t, err)
@@ -105,11 +102,12 @@ func TestBeginBlocker(t *testing.T) {
 	require.NoError(t, err)
 	// for 50 blocks, mark the validator as having not signed
 	for ; height < ((signedBlocksWindow * 2) - minSignedPerWindow + 1); height++ {
-		ctx = ctx.WithBlockHeight(height).
-			WithVoteInfos([]abci.VoteInfo{{
+		ctx = ctx.WithBlockHeight(height).WithCometInfo(comet.Info{
+			LastCommit: comet.CommitInfo{Votes: []comet.VoteInfo{{
 				Validator:   abciVal,
-				BlockIdFlag: cmtproto.BlockIDFlagAbsent,
-			}})
+				BlockIDFlag: comet.BlockIDFlagAbsent,
+			}}},
+		})
 
 		err = slashing.BeginBlocker(ctx, slashingKeeper)
 		require.NoError(t, err)