+++ x/upgrade/types/codec.go
@@ -4,17 +4,10 @@ import (
 	"github.com/cosmos/cosmos-sdk/codec/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/msgservice"
-	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types/v1beta1"
 )
 
 // RegisterInterfaces registers the interfaces types with the Interface Registry.
 func RegisterInterfaces(registry types.InterfaceRegistry) {
-	registry.RegisterImplementations(
-		(*govtypes.Content)(nil),
-		&SoftwareUpgradeProposal{},
-		&CancelSoftwareUpgradeProposal{},
-	)
-
 	registry.RegisterImplementations((*sdk.Msg)(nil),
 		&MsgSoftwareUpgrade{},
 		&MsgCancelUpgrade{},